+++ visual-dotnet/SauceLabs.Visual/Models/Browser.cs
@@ -1,0 +1,11 @@
+namespace SauceLabs.Visual.Models
+{
+    public enum Browser
+    {
+        Chrome,
+        Edge,
+        Firefox,
+        PlaywrightWebkit,
+        Safari,
+    }
+}