+++ visual-dotnet/SauceLabs.Visual/Models/OperatingSystem.cs
@@ -1,0 +1,11 @@
+namespace SauceLabs.Visual.Models
+{
+    public enum OperatingSystem
+    {
+        Android,
+        Ios,
+        Linux,
+        Macos,
+        Windows,
+    }
+}